fix: harden DINOv2 first-use acquisition
