+++ src/iPhoto/pets/model_manifest.json
@@ -19,6 +19,9 @@
     "model_name": "dinov2_vits14",
     "source_repository": "facebookresearch/dinov2",
     "source_revision": "7764ea0f912e53c92e82eb78a2a1631e92725fc8",
+    "source_tree_sha1": "2a27257b79b0633b027a21014bc9360e3c1b3f43",
+    "source_archive_url": "https://github.com/facebookresearch/dinov2/archive/7764ea0f912e53c92e82eb78a2a1631e92725fc8.zip",
+    "source_archive_max_bytes": 67108864,
     "weights_url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vits14/dinov2_vits14_pretrain.pth",
     "weights_sha256": "b938bf1bc15cd2ec0feacfe3a1bb553fe8ea9ca46a7e1d8d00217f29aef60cd9",
     "weights_size": 88283115,
